refactor(interfaces): move some interfaces into sub-folders
