+++ src/github/api/issues/github-api-issues.service.spec.ts
@@ -1,5 +1,5 @@
-import { IGithubApiIssues } from '@github/api/issues/github-api-issues.interface';
 import { GithubApiIssuesService } from '@github/api/issues/github-api-issues.service';
+import { IGithubApiIssues } from '@github/api/issues/interfaces/github-api-issues.interface';
 import { OctokitService } from '@github/octokit/octokit.service';
 import { LoggerService } from '@utils/loggers/logger.service';
 import { context } from '@actions/github';
