+++ src/github/api/issues/github-api-issues.service.ts
@@ -1,5 +1,5 @@
-import { IGithubApiIssues } from '@github/api/issues/github-api-issues.interface';
 import { GITHUB_ISSUES_PER_PAGE } from '@github/api/issues/github-issues-per-page';
+import { IGithubApiIssues } from '@github/api/issues/interfaces/github-api-issues.interface';
 import { OctokitService } from '@github/octokit/octokit.service';
 import { LoggerFormatService } from '@utils/loggers/logger-format.service';
 import { LoggerService } from '@utils/loggers/logger.service';
