+++ src/github/api/issues/interfaces/github-api-issue.interface.ts
