+++ src/github/api/issues/interfaces/github-api-issues.interface.ts
@@ -1,4 +1,4 @@
-import { IGithubApiIssuesPagination } from '@github/api/issues/github-api-issues-pagination.interface';
+import { IGithubApiIssuesPagination } from '@github/api/issues/interfaces/github-api-issues-pagination.interface';
 
 export interface IGithubApiIssues {
   readonly repository: {
