+++ src/github/api/labels/github-api-labels.service.spec.ts
@@ -1,6 +1,6 @@
-import { IGithubApiGetLabel } from '@github/api/labels/github-api-get-label.interface';
 import { IGithubApiLabel } from '@github/api/labels/github-api-label.interface';
 import { GithubApiLabelsService } from '@github/api/labels/github-api-labels.service';
+import { IGithubApiGetLabel } from '@github/api/labels/interfaces/github-api-get-label.interface';
 import { OctokitService } from '@github/octokit/octokit.service';
 import { LoggerService } from '@utils/loggers/logger.service';
 import { context } from '@actions/github';
