+++ src/github/api/labels/interfaces/github-api-get-label.interface.ts
@@ -1,4 +1,4 @@
-import { IGithubApiGetLabelPagination } from '@github/api/labels/github-api-get-label-pagination.interface';
+import { IGithubApiGetLabelPagination } from '@github/api/labels/interfaces/github-api-get-label-pagination.interface';
 
 export interface IGithubApiGetLabel {
   readonly repository: {
